+++ cloudberry/neo/app/controllers/Cloudberry.scala
@@ -39,8 +39,8 @@ class Cloudberry @Inject()(val wsClient: WSClient,
   val (asterixConn, qlGenerator) =
     config.AsterixLang match {
       case "sparksql" => (new SparkConn(url = "local[4]"), SparkSQLGenerator)
-      case "sql" => (new SQLConn(config.MySqlURL), SQLGenerator)
-      case "psql" => (new PSQLConn(config.MySqlURL), PSQLGenerator)
+      case "mysql" => (new MySQLConn(config.MySqlURL), MySQLGenerator)
+      case "postgresql" => (new PostgreSQLConn(config.MySqlURL), PostgreSQLGenerator)
       case "aql" => (new AsterixAQLConn(config.AsterixURL, wsClient), AQLGenerator)
       case "sqlpp" => (new AsterixSQLPPConn(config.AsterixURL, wsClient), SQLPPGenerator)
       case _ => throw new IllegalArgumentException(s"unknown asterixdb.lang option:${config.AsterixLang}")
