+++ cloudberry/neo/app/db/Migration_20160814.scala
@@ -1,7 +1,7 @@
 package db
 
 import edu.uci.ics.cloudberry.zion.model.datastore.IDataConn
-import edu.uci.ics.cloudberry.zion.model.impl.{DataSetInfo, SQLConn, PSQLConn, AsterixSQLPPConn}
+import edu.uci.ics.cloudberry.zion.model.impl.{DataSetInfo, MySQLConn, PostgreSQLConn, AsterixSQLPPConn}
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -12,7 +12,7 @@ private[db] class Migration_20160814() {
   //TODO it supposes to automatically register the dataset from AsterixDB
   def up(conn: IDataConn)(implicit ec: ExecutionContext): Future[Boolean] = {
     conn match {
-      case psql: PSQLConn =>
+      case psql: PostgreSQLConn =>
         conn.postControl{
           s"""
              |create table if not exists "berry.meta" (
@@ -25,7 +25,7 @@ private[db] class Migration_20160814() {
              |);
            """.stripMargin
         }
-      case sql: SQLConn =>
+      case sql: MySQLConn =>
         conn.postControl {
           s"""
              |create table if not exists `berry.meta` (
