+++ cloudberry/neo/conf/application.conf
@@ -89,11 +89,11 @@ berry.query.gap = "1 day"
 #asterixdb.url = "http://localhost:19002/aql"
 asterixdb.url = "http://localhost:19002/query/service"
 mysqldb.url = "jdbc:mysql://localhost:6603/sql?protocol=TCP&user=root"
-#acceptable values: "AQL" or "SQLPP" or "sparksql" or "SQL" or "PSQL"
+#acceptable values: "AQL" or "SQLPP" or "sparksql" or "mysql" or "postgresql"
 asterixdb.lang = SQLPP
 #asterixdb.lang = sparksql
-#asterixdb.lang = SQL
-#asterixdb.lang = PSQL
+#asterixdb.lang = mysql
+#asterixdb.lang = postgresql
 asterixdb.view.meta.name = "viewMeta"
 
 # WebSocket message size limit
