--- cloudberry/zion/src/main/scala/edu/uci/ics/cloudberry/zion/model/impl/PSQLConn.scala
@@ -1,5 +1,0 @@
-package edu.uci.ics.cloudberry.zion.model.impl
-
-import scala.concurrent.ExecutionContext
-
-class PSQLConn(url: String)(implicit ec: ExecutionContext) extends SQLConn(url) {}
