+++ cloudberry/zion/src/main/scala/edu/uci/ics/cloudberry/zion/model/impl/PostgreSQLConn.scala
@@ -1,0 +1,5 @@
+package edu.uci.ics.cloudberry.zion.model.impl
+
+import scala.concurrent.ExecutionContext
+
+class PostgreSQLConn(url: String)(implicit ec: ExecutionContext) extends MySQLConn(url) {}
